requirements.txt
fastapi[standard]~=0.114.2
pymongo[srv]~=4.8.0
pydantic~=2.9.1
python-dotenv~=1.0.1
motor==3.5.1
